+++ src/lbm_bnd_interface.h
@@ -11,7 +11,7 @@ class LBMBndInterface {
   auto operator=(const LBMBndInterface&) -> LBMBndInterface& = delete;
   auto operator=(LBMBndInterface&&) -> LBMBndInterface& = delete;
 
-  virtual void init()                                                    = 0;
+//  virtual void init()                                                    = 0;
   virtual void preApply(const std::function<GDouble&(GInt, GInt)>& f, const std::function<GDouble&(GInt, GInt)>& fold,
                         const std::function<GDouble&(GInt, GInt)>& vars) = 0;
   virtual void apply(const std::function<GDouble&(GInt, GInt)>& f, const std::function<GDouble&(GInt, GInt)>& fold,