+++ CHANGELOG.md
@@ -13,6 +13,10 @@
 - capalilary simulation
 - contact angle
 - output analytical solution if wanted
+- check in periodic boundaries if all cells have been linked
+- get boundary const access to grid information
+- pass all surfaces to the boundary manager
+- store reference to the surfaces in each boundary
 
 # V0.0.1
 ## Planned: